services/ari-service/requirements.txt
aiohttp==3.10.5
websockets==15.0.1
paho-mqtt==1.6.1
